split common utilities
